backend/requirements.txt
Flask
flask-cors
mysql-connector-python
PyJWT
werkzeug
